Большой рефакторинг, добавлен период (в разработке)

+++ src/daemon/main.rs
@@ -2,26 +2,27 @@ use std::io::{self, BufReader, BufWriter};
 use std::error::Error;
 use std::os::fd::AsFd;
 use std::os::unix::net::{UnixListener, UnixStream};
+use nix::errno::Errno;
 use nix::sys::signal::{Signal, SigSet};
 use nix::sys::signalfd::{SignalFd, SfdFlags};
-use nix::poll::{PollFd, PollFlags, PollTimeout, poll};
+use nix::poll::{self, PollFd, PollFlags, PollTimeout};
 
-use ywpm::server::backend::stop_backend;
-use ywpm::server::state::State;
-use ywpm::server::args_parser::parse_args;
-use ywpm::reader::read_string_vec;
-use ywpm::writer::write_response;
-use ywpm::util::{get_socket_path, get_config_path};
+use ywpm::daemon::arg_parsing::arg_parser;
+use ywpm::daemon::backend;
+use ywpm::daemon::state::State;
+use ywpm::reader;
+use ywpm::writer;
+use ywpm::util;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let mut state = State::read_or_create_empty(get_config_path())?;
+    let mut state = State::read_or_create_empty(util::get_config_path())?;
     state.start_backend()?;
     main_loop(state)
 }
 
 
 fn main_loop(mut state: State) -> Result<(), Box<dyn Error>> {
-    let socket_path = get_socket_path();
+    let socket_path = util::get_socket_path();
 
     let listener = UnixListener::bind(&socket_path)?;
     listener.set_nonblocking(true)?;
@@ -39,7 +40,7 @@ fn main_loop(mut state: State) -> Result<(), Box<dyn Error>> {
             PollFd::new(signal_fd.as_fd(), PollFlags::POLLIN),
         ];
 
-        match poll(&mut poll_fds, PollTimeout::try_from(timeout_ms)?) {
+        match poll::poll(&mut poll_fds, PollTimeout::try_from(timeout_ms)?) {
             Ok(0) => {
                 println!("Timeout reached! Changing wallpaper...");
                 
@@ -47,29 +48,29 @@ fn main_loop(mut state: State) -> Result<(), Box<dyn Error>> {
             }
 
             Ok(_) => {
-                if let Some(revents) = poll_fds[0].revents() {
-                    if revents.contains(PollFlags::POLLIN) {
-                        match listener.accept() {
-                            Ok((stream, _)) => handle_client(stream, &mut state)?,
-                            Err(err) if err.kind() == io::ErrorKind::WouldBlock => (),
-                            Err(err) => eprintln!("Accept error: {err}"),
-                        }
+                if let Some(revents) = poll_fds[0].revents() &&
+                    revents.contains(PollFlags::POLLIN) {
+                    
+                    match listener.accept() {
+                        Ok((stream, _)) => handle_client(stream, &mut state)?,
+                        Err(err) if err.kind() == io::ErrorKind::WouldBlock => (),
+                        Err(err) => eprintln!("Accept error: {err}"),
                     }
                 }
 
-                if let Some(revents) = poll_fds[1].revents() {
-                    if revents.contains(PollFlags::POLLIN) {
-                        println!("\nReceived shutdown signal! Exiting gracefully...");
+                if let Some(revents) = poll_fds[1].revents() &&
+                    revents.contains(PollFlags::POLLIN) {
 
-                        let _ = std::fs::remove_file(&socket_path);
+                    println!("\nReceived shutdown signal! Exiting gracefully...");
 
-                        if let Err(err) = state.write_to(get_config_path()) {
-                            eprintln!("Error while writing config: {}", err.to_string());
-                        }
-                        
-                        stop_backend();
-                        break; 
+                    let _ = std::fs::remove_file(&socket_path);
+
+                    if let Err(err) = state.write_to(util::get_config_path()) {
+                        eprintln!("Error while writing config: {}", err.to_string());
                     }
+                    
+                    backend::stop();
+                    break; 
                 }
             }
 
@@ -90,7 +91,7 @@ fn main_loop(mut state: State) -> Result<(), Box<dyn Error>> {
 }
 
 
-fn get_signal_fd() -> Result<SignalFd, Box<dyn Error>> {
+fn get_signal_fd() -> Result<SignalFd, Errno> {
     let mut mask = SigSet::empty();
     mask.add(Signal::SIGINT);
     mask.add(Signal::SIGTERM);
@@ -101,19 +102,19 @@ fn get_signal_fd() -> Result<SignalFd, Box<dyn Error>> {
 
 
 fn handle_client(mut stream: UnixStream, state: &mut State) -> Result<(), Box<dyn Error>> {
-    let args = read_string_vec(&mut BufReader::new(&mut stream))?;
+    let args = reader::read_string_vec(&mut BufReader::new(&mut stream))?;
 
     let mut writer = BufWriter::new(&mut stream);
     
-    match parse_args(&args) {
+    match arg_parser::parse_args(&args) {
         Ok(action) => {
             match action.perform(args[0].as_str(), state) {
-                Ok(message) => write_response(&mut writer, true, &message)?,
-                Err(err)    => write_response(&mut writer, false, err.message())?,
+                Ok(message) => writer::write_response(&mut writer, true, &message)?,
+                Err(err)    => writer::write_response(&mut writer, false, err.message())?,
             }
         },
 
-        Err(err) => write_response(&mut writer, false, err.message())?,
+        Err(err) => writer::write_response(&mut writer, false, err.message())?,
     }
 
     Ok(())
